server/pyproject.toml
[tool.poetry]
name = "ohserver"
version = "0.1.0"
description = "Queue@Cornell backend server for the final web application"
authors = ["Magd Bayoumi <mb2363@cornell.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
typing-extensions = "^3.7.4"
requests = "^2.23.0"
flask = "^1.1.1"
SQLAlchemy = "^1.3.15"
graphql-core = ">=2.1,<3"
graphql-server-core = "^1.2.0"
graphene = "^2.1.8"
graphql-relay = "^2.0.1"
flask-graphql = "^2.0.1"
Flask-SQLAlchemy = "^2.4.1"
graphene-sqlalchemy = "^2.2.2"
flask_cors = "^3.0.8"
gevent = "^1.4.0"
flask-sockets = "^0.2.1"
graphql-ws = "^0.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
pytest-cov = "^2.8"
pytest-mock = "^1.12"
xdoctest = "^0.10.3"
coverage = "^4.5"
flake8 = "^3.7"
flake8-docstrings = "^1.5"
flake8-rst-docstrings = "^0.0.12"
darglint = "^1.1"
isort = {version = "^4.3", extras = ["pyproject"]}
seed-isort-config = "^1.9"
pre-commit = "^1.20"
black = {version = "^18.3-alpha.0", allow-prereleases = true}
sphinx = "^2.2"
sphinx-rtd-theme = "^0.4.3"
sphinx_pdj_theme = "^0.1.5"
pallets_sphinx_themes = "^1.2.3"
sphinxcontrib-openapi = "^0.6.0"
keyring = "^21.0.0"
artifacts-keyring = "^0.2.8-rc.0"
tox = "^3.14.3"
pytorch_sphinx_theme = "^0.0.19"
sphinxcontrib-katex = "^0.6.0"
tqdm = "^4.43.0"
mypy = "^0.770"
pylint = "^2.4.4"

[tool.poetry.extras]
doc = ["sphinx", "sphinx_rtd_theme"]

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = true
envlist = py36,py37,py38
[testenv]
whitelist_externals = poetry
commands =
    poetry install -vvv
    poetry run python -m pytest
"""

[tool.isort]
known_first_party = 'server'
known_third_party = ["matplotlib", "numpy", "pallets_sphinx_themes"]
multi_line_output = 3
lines_after_imports = 2
force_grid_wrap = 0
combine_as_imports = true
include_trailing_comma = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
